Fixed routes name and displayed ratin on video list

+++ database/migrations/2020_05_26_235825_create_videos_table.php
@@ -21,7 +21,7 @@ class CreateVideosTable extends Migration
             $table->string('description')->nullable();
             $table->string('category');
             $table->string('poster_id')->nullable();
-            $table->integer('rating')->nullable();
+            $table->float('rating')->nullable();
             $table->bigInteger('views')->nullable()->default(0);
             $table->timestamps();
         });
